+++ zcord/bot.py
@@ -4,23 +4,37 @@ import asyncio
 import contextlib
 import logging
 from importlib.metadata import version
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 import aiohttp
 
+from zcord import enums
+from zcord._logging import setup_logging
 from zcord.gateway import Gateway
 from zcord.models.channel import Channel
+from zcord.models.guild import Guild
 from zcord.models.message import Message
 from zcord.state import ConnectionState
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from zcord import bitfields
     from zcord.models.application import Application
-    from zcord.models.guild import Guild
+    from zcord.models.base import Model
     from zcord.models.snowflake import Snowflake
     from zcord.models.user import User
 
 log = logging.getLogger(__name__)
+
+_EVENT_MODELS: dict[str, type[Model]] = {
+    enums.GatewayEvent.GUILD_CREATE.value: Guild,
+    enums.GatewayEvent.MESSAGE_CREATE.value: Message,
+}
+
+_UPDATE_EVENTS: dict[str, tuple[type[Model], str]] = {
+    enums.GatewayEvent.GUILD_UPDATE.value: (Guild, "_guilds")
+}
 
 
 class Bot:
@@ -45,7 +59,10 @@ class Bot:
         self._state = ConnectionState(token)
         if intents is not None:
             self._state._gateway = Gateway(
-                http=self._state._http, token=token, intents=intents
+                http=self._state._http,
+                token=token,
+                intents=intents,
+                dispatch=self._dispatch,
             )
         else:
             log.warning(
@@ -53,6 +70,9 @@ class Bot:
             )
         Message._state = self._state
         Channel._state = self._state
+
+        self._events: dict[str, list[tuple[Callable[..., Any], bool]]] = {}
+        self._tasks: set[asyncio.Task] = set()
 
     async def __aenter__(self) -> Bot:
         return self
@@ -76,13 +96,83 @@ class Bot:
 
         async def _connect() -> None:
             if self._state._gateway is not None:
-                await self._state._gateway.run()
-            done.set()
+                try:
+                    await self._state._gateway.run()
+                finally:
+                    done.set()
 
         task = asyncio.create_task(_connect())
         with contextlib.suppress(KeyboardInterrupt):
             await done.wait()
         task.cancel()
+
+    def run(self) -> None:
+        """
+        Run the bot loop.
+        """
+        if not logging.getLogger("zcord").handlers:
+            setup_logging()
+        asyncio.run(self.start())
+
+    def on(
+        self, event: str | enums.GatewayEvent, callback: Callable[..., Any]
+    ) -> None:
+        """
+        Register a persistent event listener.
+        """
+        self._events.setdefault(str(event), []).append((callback, False))
+
+    def once(
+        self, event: str | enums.GatewayEvent, callback: Callable[..., Any]
+    ) -> None:
+        """
+        Register a one-time event listener.
+        """
+        self._events.setdefault(str(event), []).append((callback, True))
+
+    def _dispatch(self, event: str, *args: Any) -> None:
+        """
+        Dispatch an event to all registered listeners.
+        """
+        listeners = self._events.get(event, [])
+        self._events[event] = [
+            callback for callback in listeners if not callback[1]
+        ]  # keep persistent listeners
+        for callback, _ in listeners:
+            data = args[0] if args else None
+            # Update event will have 2 args: old and new
+            if data and event in _UPDATE_EVENTS:
+                model, cache_attr = _UPDATE_EVENTS[event]
+                obj_id = int(data["id"])
+                old = getattr(self._state, cache_attr, {}).get(obj_id)
+                new = model._from_payload(data)
+                try:
+                    maybe_coro = callback(old, new)
+                except Exception:
+                    log.exception("Failed to dispatch event %s", event)
+                    continue
+            # Some other events will have 1 arg: the object
+            elif data and event in _EVENT_MODELS:
+                try:
+                    maybe_coro = callback(
+                        _EVENT_MODELS[event]._from_payload(data)
+                    )
+                except Exception:
+                    log.exception("Failed to dispatch event %s", event)
+                    continue
+            else:
+                try:
+                    maybe_coro = callback(*args) if args else callback()
+                except Exception:
+                    log.exception("Failed to dispatch event %s", event)
+                    continue
+            if asyncio.iscoroutine(maybe_coro):
+                task = asyncio.create_task(maybe_coro)
+                self._tasks.add(task)
+                task.add_done_callback(self._tasks.discard)
+
+        if args:
+            self._state._update_cache(event, args[0])
 
     async def fetch_current_application(self) -> Application:
         """