+++ zcord/enums/application.py
@@ -14,3 +14,7 @@ class EventWebhookStatus(IntEnum):
     | `ENABLED` | `2` |
     | `DISABLED_BY_DISCORD` | `3` |
     """
+
+    DISABLED = 1
+    ENABLED = 2
+    DISABLED_BY_DISCORD = 3